+++ examples/mtext.py
@@ -4,13 +4,15 @@
 # License: MIT License
 import ezdxf
 
-dwg = ezdxf.new('ac1015')
-modelspace = dwg.modelspace()
+dwg = ezdxf.new('R2007')
+msp = dwg.modelspace()
 attribs = {
     'char_height': 0.7,
     'width': 5.0,
 }
-modelspace.add_mtext("This is a long MTEXT line with line wrapping!", attribs)
+msp.add_line((-10, -1), (10, -2))
+mtext = msp.add_mtext("This is a long MTEXT line with line wrapping!", attribs)
+mtext.set_bg_color((108, 204, 193))
 
 filename = 'mtext.dxf'
 dwg.saveas(filename)